# openapi.yaml
openapi: 3.0.3
info:
  title: MonPetitBiz WhatsApp Bot API
  description: |
    A WhatsApp bot API for micro-businesses in Africa to record transactions, 
    track inventory, and generate financial reports through natural language conversations.
    
    ## Features
    - WhatsApp webhook integration
    - Natural language processing for business commands
    - Transaction recording (sales & expenses)
    - Inventory management
    - Financial report generation
    - PDF report delivery via WhatsApp
    - Comprehensive health monitoring
    
    ## Authentication
    Most endpoints require JWT authentication. Use the `/auth/send-otp` and 
    `/auth/verify-otp` endpoints to obtain an access token.
    
  version: 1.0.0
  contact:
    name: MonPetitBiz Support
    email: support@monpetitbiz.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.monpetitbiz.com
    description: Production server
...
paths:
...
  /whatsapp/twilio/webhook:
    post:
      tags:
        - WhatsApp
      summary: Receive Twilio WhatsApp messages (Primary)
      description: |
        Primary webhook endpoint for processing incoming WhatsApp messages from Twilio Programmable Messaging API.
        This endpoint handles webhook signature verification and processes messages using Twilio's native format.
        
        **Webhook Configuration:**
        - URL: `https://yourdomain.com/whatsapp/twilio/webhook`
        - Method: POST
        - Content-Type: application/x-www-form-urlencoded
        - Signature verification: X-Twilio-Signature header
        
        **Message Processing:**
        - Extracts phone numbers from Twilio format (whatsapp:+number)
        - Uses MessageSid for deduplication
        - Processes business commands (sales, expenses, stock, reports)
        - Handles user registration and authentication
      parameters:
        - name: X-Twilio-Signature
          in: header
          description: Twilio webhook signature for payload verification
          required: true
          schema:
            type: string
            example: "sha256=abc123..."
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioWebhookPayload'
            examples:
              sale_command:
                summary: Sale command from Twilio
                value:
                  MessageSid: "SM1234567890abcdef1234567890abcdef"
                  AccountSid: "AC1234567890abcdef1234567890abcdef"
                  From: "whatsapp:+221123456789"
                  To: "whatsapp:+14155238886"
                  Body: "vente pain 1500"
                  ProfileName: "Fabrice Ilboudo"
                  WaId: "221123456789"
                  SmsStatus: "received"
              expense_command:
                summary: Expense command from Twilio
                value:
                  MessageSid: "SM9876543210fedcba9876543210fedcba"
                  AccountSid: "AC1234567890abcdef1234567890abcdef"
                  From: "whatsapp:+221987654321"
                  To: "whatsapp:+14155238886"
                  Body: "dépense 2000 marchandise"
                  ProfileName: "Moussa Sow"
                  WaId: "221987654321"
                  SmsStatus: "received"
              registration:
                summary: User registration
                value:
                  MessageSid: "SM5555555555abcdef5555555555abcdef"
                  AccountSid: "AC1234567890abcdef1234567890abcdef"
                  From: "whatsapp:+221555666777"
                  To: "whatsapp:+14155238886"
                  Body: "bonjour"
                  ProfileName: "Nouveau Utilisateur"
                  WaId: "221555666777"
                  SmsStatus: "received"
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message processed successfully"
                  messageId:
                    type: string
                    example: "SM1234567890abcdef1234567890abcdef"
                    description: "Twilio MessageSid for tracking"
        '400':
          description: Invalid webhook payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_payload:
                  summary: Invalid payload format
                  value:
                    statusCode: 400
                    message: "Invalid webhook payload format"
                    error: "Bad Request"
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_signature:
                  summary: Signature verification failed
                  value:
                    statusCode: 401
                    message: "Invalid webhook signature"
                    error: "Unauthorized"
        '500':
          description: Internal server error during message processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /whatsapp/twilio:
    post:
      tags:
        - WhatsApp
      summary: Receive Twilio WhatsApp messages (Legacy)
      description: |
        Legacy webhook endpoint for Twilio WhatsApp messages. 
        **Deprecated**: Use `/whatsapp/twilio/webhook` instead.
        
        This endpoint is maintained for backward compatibility but may be removed in future versions.
      deprecated: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioWebhookPayload'
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message processed successfully"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TwilioWebhookPayload:
      type: object
      description: Twilio Programmable Messaging webhook payload for WhatsApp messages
      required:
        - MessageSid
        - AccountSid
        - From
        - To
        - Body
        - WaId
      properties:
        MessageSid:
          type: string
          description: Unique identifier for the message
          example: "SM1234567890abcdef1234567890abcdef"
          pattern: '^SM[a-f0-9]{32}$'
        AccountSid:
          type: string
          description: Twilio Account SID
          example: "AC1234567890abcdef1234567890abcdef"
          pattern: '^AC[a-f0-9]{32}$'
        From:
          type: string
          description: Sender's WhatsApp number in Twilio format
          example: "whatsapp:+221123456789"
          pattern: '^whatsapp:\+[1-9]\d{1,14}$'
        To:
          type: string
          description: Recipient's WhatsApp number in Twilio format
          example: "whatsapp:+14155238886"
          pattern: '^whatsapp:\+[1-9]\d{1,14}$'
        Body:
          type: string
          description: Message content
          example: "vente pain 1500"
          maxLength: 1600
        ProfileName:
          type: string
          description: WhatsApp profile name of the sender (optional)
          example: "Fabrice Ilboudo"
          maxLength: 256
        WaId:
          type: string
          description: WhatsApp ID (phone number without country code formatting)
          example: "221123456789"
          pattern: '^\d{10,15}$'
        SmsStatus:
          type: string
          description: Message status from Twilio
          enum: [received, sent, delivered, failed, undelivered]
          example: "received"
        NumMedia:
          type: string
          description: Number of media attachments (optional)
          example: "0"
          pattern: '^\d+$'
        MediaUrl0:
          type: string
          description: URL of first media attachment (if NumMedia > 0)
          format: uri
          example: "https://api.twilio.com/2010-04-01/Accounts/.../Messages/.../Media/..."
        MediaContentType0:
          type: string
          description: Content type of first media attachment
          example: "image/jpeg"
        MessageStatus:
          type: string
          description: Current status of the message
          enum: [accepted, queued, sending, sent, failed, delivered, undelivered, receiving, received]
          example: "received"
        ErrorCode:
          type: string
          description: Error code if message failed (optional)
          example: "30008"
        ErrorMessage:
          type: string
          description: Error message if message failed (optional)
          example: "Unknown error"
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: "Invalid request"
        error:
          type: string
          example: "Bad Request"
        timestamp:
          type: string
          format: date-time
...